Deduplicate TestServer tests
