+++ gotham/src/tls/test.rs
@@ -270,82 +270,3 @@ impl Service<Uri> for TestConnect {
         .boxed()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    use http::header::CONTENT_TYPE;
-    use hyper::header::CONTENT_LENGTH;
-    use hyper::StatusCode;
-
-    use crate::test::helper::TestHandler;
-
-    #[test]
-    fn serves_requests() {
-        let ticks = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_secs();
-
-        let test_server = TestServer::new(TestHandler::from(format!("time: {}", ticks))).unwrap();
-
-        let response = test_server
-            .client()
-            .get("https://example.com/")
-            .perform()
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-        let buf = response.read_utf8_body().unwrap();
-        assert_eq!(buf, format!("time: {}", ticks));
-    }
-
-    #[test]
-    fn times_out() {
-        // sadly it seems near impossible to use `tokio::time::advance` to test this
-        let timeout = Duration::from_secs(1);
-        let test_server =
-            TestServer::with_timeout(TestHandler::default(), timeout.as_secs()).unwrap();
-        let result = test_server
-            .client()
-            .get("http://localhost/timeout")
-            .perform();
-        assert!(result.unwrap_err().to_string().contains("timed out"));
-    }
-
-    #[test]
-    fn async_echo() {
-        let server = TestServer::new(TestHandler::default()).unwrap();
-
-        let client = server.client();
-        let data = "This text should get reflected back to us. Even this fancy piece of unicode: \
-                    \u{3044}\u{308d}\u{306f}\u{306b}\u{307b}";
-
-        let res = client
-            .post("https://example.com/echo", data, mime::TEXT_PLAIN)
-            .perform()
-            .expect("request successful");
-
-        assert_eq!(res.status(), StatusCode::OK);
-
-        {
-            let mime = res.headers().get(CONTENT_TYPE).expect("ContentType");
-            assert_eq!(mime, mime::TEXT_PLAIN.as_ref());
-        }
-
-        let content_length = {
-            let content_length = res.headers().get(CONTENT_LENGTH).expect("ContentLength");
-            assert_eq!(content_length, &format!("{}", data.as_bytes().len()));
-            content_length.clone()
-        };
-
-        let buf =
-            String::from_utf8(res.read_body().expect("readable response")).expect("UTF8 response");
-
-        assert_eq!(content_length, &format!("{}", buf.len()));
-        assert_eq!(data, &buf);
-    }
-}
